refactor: quit separating 'skipstage' package
