+++ pkg/app/piped/controller/skipstage.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package skipstage
+package controller
 
 import (
 	"context"
