+++ pkg/app/piped/controller/skipstage_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package skipstage
+package controller
 
 import (
 	"testing"
